+++ zia/services/adminuserrolemgmt/roles/adminroles.go
@@ -1,4 +1,4 @@
-package adminuserrolemgmt
+package roles
 
 import (
 	"fmt"