+++ zia/services/adminuserrolemgmt/roles/service.go
@@ -1,4 +1,4 @@
-package usermanagement
+package roles
 
 import (
 	"github.com/zscaler/zscaler-sdk-go/v2/zia"