+++ zia/services/usermanagement/users/users_test.go
@@ -1,4 +1,4 @@
-package usermanagement
+package users
 
 import (
 	"log"