+++ zia/services/intermediatecacertificates/intermediatecacertificates_test.go
@@ -1,9 +1,12 @@
 package intermediatecacertificates
 
 import (
+	"strings"
 	"testing"
 
 	"github.com/zscaler/zscaler-sdk-go/v2/tests"
+	"golang.org/x/text/cases"
+	"golang.org/x/text/language"
 )
 
 func TestIntermediateCertificate_data(t *testing.T) {
@@ -65,4 +68,38 @@ func TestIntermediateCertificate_data(t *testing.T) {
 	}
 	t.Logf("Successfully retrieved ready-to-use intermediate certificate: %s", readyToUseCerts[0].Name)
 
+}
+
+func TestCaseSensitivityOfGetByName(t *testing.T) {
+	client, err := tests.NewZiaClient()
+	if err != nil {
+		t.Errorf("Error creating client: %v", err)
+		return
+	}
+
+	service := New(client)
+
+	// Assuming a certificate with the name "Zscaler Intermediate CA Certificate" exists
+	knownName := "Zscaler Intermediate CA Certificate"
+
+	// Case variations to test
+	variations := []string{
+		strings.ToUpper(knownName),
+		strings.ToLower(knownName),
+		cases.Title(language.English).String(knownName),
+	}
+
+	for _, variation := range variations {
+		t.Logf("Attempting to retrieve group with name variation: %s", variation)
+		certificate, err := service.GetByName(variation)
+		if err != nil {
+			t.Errorf("Error getting certificate with name variation '%s': %v", variation, err)
+			continue
+		}
+
+		// Check if the certificate's actual name matches the known name
+		if certificate.Name != knownName {
+			t.Errorf("Expected certificate name to be '%s' for variation '%s', but got '%s'", knownName, variation, certificate.Name)
+		}
+	}
 }