+++ CHANGELOG.md
@@ -5,9 +5,11 @@
 ## Notes
 - Golang: **v1.19**
 
-### Fixes
+### Enhacements
 
 - [PR #202](https://github.com/zscaler/zscaler-sdk-go/pull/202) Added support to 🆕 ZIA Cloud Browser Isolation Profile endpoint ``/browserIsolation/profiles``
+
+- [PR #202](https://github.com/zscaler/zscaler-sdk-go/pull/202) Added `cbiProfile` feature to ZIA `url filtering policy` resource
 
 # 2.2.2 (December 10, 2023)